requirements/base.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
psycopg==3.2.6
gunicorn==23.0.0
python-dotenv==1.1.0
flask-restx==1.3.0
Flask-Migrate==4.1.0
bcrypt==4.3.0
PyJWT==2.10.1
parse==1.20.2
cryptography==44.0.2